Move mesh and skinned mesh to renderer
