+++ editor/src/quoll/editor/scene/renderer/EditorRenderer.cpp
@@ -1,8 +1,8 @@
 #include "quoll/core/Base.h"
 #include "quoll/core/Engine.h"
-#include "quoll/scene/Mesh.h"
+#include "quoll/renderer/Mesh.h"
 #include "quoll/scene/Sprite.h"
-#include "quoll/scene/SkinnedMesh.h"
+#include "quoll/renderer/SkinnedMesh.h"
 #include "quoll/skeleton/Skeleton.h"
 #include "quoll/text/Text.h"
 #include "quoll/renderer/MeshVertexLayout.h"
