+++ editor/src/quoll/editor/scene/ui/SceneHierarchyPanel.cpp
@@ -2,8 +2,8 @@
 #include "quoll/core/Name.h"
 #include "quoll/scene/EnvironmentSkybox.h"
 #include "quoll/scene/Camera.h"
-#include "quoll/scene/Mesh.h"
-#include "quoll/scene/SkinnedMesh.h"
+#include "quoll/renderer/Mesh.h"
+#include "quoll/renderer/SkinnedMesh.h"
 #include "quoll/scene/Sprite.h"
 #include "quoll/scene/LocalTransform.h"
 #include "quoll/scene/Parent.h"
