+++ engine/src/quoll/entity/EntityDatabase.cpp
@@ -3,8 +3,8 @@
 #include "quoll/core/Name.h"
 #include "quoll/core/Id.h"
 #include "quoll/core/Delete.h"
-#include "quoll/scene/Mesh.h"
-#include "quoll/scene/SkinnedMesh.h"
+#include "quoll/renderer/Mesh.h"
+#include "quoll/renderer/SkinnedMesh.h"
 #include "quoll/scene/DirectionalLight.h"
 #include "quoll/scene/PointLight.h"
 #include "quoll/scene/CascadedShadowMap.h"
