+++ engine/src/quoll/physics/PhysicsSystem.cpp
@@ -5,10 +5,6 @@
 
 namespace quoll {
 
-PhysicsSystem PhysicsSystem::createPhysxBackend() {
-  return std::move(PhysicsSystem(new PhysxBackend));
-}
-
 PhysicsSystem::PhysicsSystem(PhysicsBackend *backend) : mBackend(backend) {}
 
 } // namespace quoll
