+++ engine/src/quoll/renderer/MeshSerializer.cpp
