+++ engine/src/quoll/renderer/SceneRenderer.cpp
@@ -1,8 +1,6 @@
 #include "quoll/core/Base.h"
 #include "quoll/core/Engine.h"
 #include "quoll/scene/EnvironmentSkybox.h"
-#include "quoll/scene/Mesh.h"
-#include "quoll/scene/SkinnedMesh.h"
 #include "quoll/scene/EnvironmentLighting.h"
 #include "quoll/scene/Sprite.h"
 #include "quoll/text/Text.h"
@@ -12,6 +10,8 @@
 #include "StandardPushConstants.h"
 #include "BindlessDrawParameters.h"
 #include "MeshVertexLayout.h"
+#include "Mesh.h"
+#include "SkinnedMesh.h"
 #include "MeshRenderUtils.h"
 
 namespace quoll {
